+++ src/lib/auth/hooks/useAuth.tsx
